Move the packages around for better organization as per the suggestions.
1. New bootstrap package that users can use.
2. Move insecure creds builder to google/creds.
3. Move registration to google/ and google/insecure.
